// src/app/shared/models/palabrasReservadas.ts
export interface PalabrasReservadas {
    idPalabraReservada?: number;
    nombre?:             string;
    orden?:              number;
    padre?:              number;
    lista?:              number;
    categoria?:          string;
    tiempo?:             number;
    usuarioCreador?:     string;
    usuarioModificador?: string;
    fechaCreacion?:      Date;
    fechaModificacion?:  Date;
}